git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/RobWork/trunk@581 494c09c7-a67d-49d4-9938-151fb1965be8

+++ src/rwlibs/drawable/RenderTriSoup.hpp
@@ -76,9 +76,9 @@ namespace rwlibs { namespace drawable {
          */
         void addFaces(
             const std::vector<rw::geometry::Face<float> >& faces,
-            float r,
-            float g,
-            float b);
+            double r,
+            double g,
+            double b);
 
         /**
          * @brief Clears the list of faces
